Move status enumerations to common
